third_party/blink/web_tests/platform/mac-mac15-arm64/fast/webgl/webgl-context-attributes-default-value-expected.txt

CONSOLE ERROR: Uncaught Unable to fetch WebGL rendering context for Canvas
Testing default value:
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
